bugs: picking first scenario, emptyStore
It wasn't picking the first scenario on load because simu.sliderTypes was not defined yet.  Put code in reset2 to click on first scenario to set it as default if it is there.
Fixed the emptyStore routine, separated from the reset Store.
fixed switching scenarios to change sliders displayed if changing INV methods.
Also saved a bunch of HTML files so they adjusted their style....