Legends update and...
Made a box for the dot and text of legend.   Made sure the baseline of the dot and baseline of the text aligned.   four items in CSS.

Also set the scenarios that when one is unselected that arrow keys will reselect it later rather than starting at the top or the bottom.